+++ build.sh
@@ -14,25 +14,22 @@ set -eu
 set -o pipefail
 
 FILES=(
-    content_scripts/autohalt-common.js
+    content_scripts/autohalt.js
     content_scripts/autohalt-facebook.js
     content_scripts/autohalt-gfycat.js
     content_scripts/autohalt-twitch.js
     content_scripts/autohalt-youtube.js
-    settings/settings.html
-    settings/settings.js
 )
 FF_FILES=(
     "${FILES[@]}"
     manifest.json
     icons/autohalt.svg
+    settings/settings.html
+    settings/settings.js
 )
 CH_FILES=(
     "${FILES[@]}"
 )
-
-# From https://github.com/mozilla/webextension-polyfill/
-PF_URL='https://unpkg.com/webextension-polyfill@0.6.0/dist/browser-polyfill.js'
 
 name="$(jq -r '.name|ascii_downcase' <manifest.json)"
 version="$(jq -r '.version' <manifest.json)"
@@ -58,20 +55,16 @@ rsync -R "${CH_FILES[@]}" "${ch_build_dir}"
 ## Apply manifest changes for chrome
 jq -f chrome.jq <manifest.json >"${ch_build_dir}/manifest.json"
 CH_FILES+=(manifest.json)
-## Add browser-polyfill.js
-download "${PF_URL}" "${ch_build_dir}/browser-polyfill.js"
-### Remove source map comment, see
-### https://bugs.chromium.org/p/chromium/issues/detail?id=212374
-sed -i.bak '\://# sourceMappingURL=.*:d' "${ch_build_dir}/browser-polyfill.js"
-CH_FILES+=(browser-polyfill.js)
-sed -e 's/\(\s*\)<!-- \(.*browser-polyfill.js.*\) -->/\1\2/g' <settings/settings.html >"${ch_build_dir}/settings/settings.html"
+## Update script configuration
+sed -i.bak 's/const useSettings = true;/const useSettings = false;/g' "${ch_build_dir}/content_scripts/autohalt.js"
 ## Generate png icons
 mkdir "${ch_build_dir}/icons/"
 sizes="$(jq -r '.icons|keys[]' <"${ch_build_dir}/manifest.json")"
 for p in ${sizes}; do
-    convert -density 2400 -resize "${p}x${p}" -background none icons/autohalt.svg "${ch_build_dir}/icons/autohalt-${p}.png"
+    convert -density 2400 -resize "${p}x${p}" -background none icons/autohalt.svg "${ch_build_dir}/icons/autohalt-${p}.png" &
     CH_FILES+=("icons/autohalt-${p}.png")
 done
+wait
 ## Package
 (cd "${ch_build_dir}" && zip "${vname}-ch.zip" "${CH_FILES[@]}")
 mv -i "${ch_build_dir}/${vname}-ch.zip" "artifacts/${vname}-ch.zip"